requirements.txt
arrow==1.3.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.1.7
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-filter==25.1
django-import-export==4.3.7
django-redis==5.4.0
django-rq==2.10.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rq==1.16.2
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tablib==3.8.0
times==0.7
types-python-dateutil==2.9.0.20241206
tzdata==2025.2
urllib3==2.3.0
